backend/requirements.txt
django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pillow==9.0.1